pyproject.toml
[build-system]
requires = ["hatchling>=1.0"]
build-backend = "hatchling.build"

[project]
name = "gbxcule-learning-lab"
version = "0.0.0"
description = "GPU-native many-env Game Boy runtime (Warp→CUDA) + benchmark/verification harness"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "libcst",
    "matplotlib",
    "numpy",
    "pillow",
    "pyyaml",
    "pyboy>=2.6.1",
    "warp-lang>=1.5.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-xdist",
    "hypothesis",
    "ruff",
    "pyright",
]
rl = [
    "torch==2.9.1",
    "torchvision==0.24.1",
]
puffer = [
    "pufferlib @ git+https://github.com/thatguy11325/PufferLib.git@1.0",
    "gymnasium",
    "psutil",
    "filelock",
]

[tool.uv]
# Resolve lock for the platforms we actually support in dev.
# This improves lock performance and avoids unsatisfiable branches.
environments = ["sys_platform == 'darwin'", "sys_platform == 'linux'"]

[tool.uv.sources]
# Force CUDA 13 wheels on Linux (DGX Spark is aarch64).
warp-lang = [
    { url = "https://github.com/NVIDIA/warp/releases/download/v1.11.0/warp_lang-1.11.0+cu13-py3-none-manylinux_2_34_aarch64.whl", marker = "sys_platform == 'linux' and platform_machine == 'aarch64'" },
    { url = "https://github.com/NVIDIA/warp/releases/download/v1.11.0/warp_lang-1.11.0+cu13-py3-none-manylinux_2_28_x86_64.whl", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'" },
]
torch = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
]
torchvision = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.hatch.build.targets.wheel]
# src-layout packaging: ship the package from src/gbxcule
packages = ["src/gbxcule"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
# Tools scripts intentionally modify sys.path before imports
"tools/*.py" = ["E402"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:Using SDL2 binaries from pysdl2-dll:UserWarning",
    "ignore:(?s)Found GPU0 NVIDIA GB10.*supported by this version of PyTorch.*:UserWarning:torch\\.cuda",
    "ignore::UserWarning:torch\\.cuda",
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
include = ["src", "tests", "bench", "tools"]
# Exclude template files (transformed at runtime via libcst) and Warp kernel DSL
exclude = [
    "src/gbxcule/kernels/cpu_templates",
    "src/gbxcule/kernels/ppu_step.py",
    "src/gbxcule/kernels/ppu_render_downsampled.py",
]
stubPath = "typings"
# Allow tests to import 'harness' from bench/
extraPaths = ["bench"]
# Allow bytearray where bytes is expected (Python runtime allows this)
disableBytesTypePromotions = false
# Don't require stubs for our own internal modules
reportMissingTypeStubs = false
# Enforce strict mode for shared contract modules
strict = ["src/gbxcule/backends/common.py"]
